delta_generator: Convert ABORT into an exit(1).
When verifying a signature in delta_generator, a workflow rarely used,
the program will abort if the signature is not valid. The unittests
exercise this flow by executing "delta_generator" with invalid keys
which causes it to abort during normal test runs. This creates lots of
unnecessary crash reports.

This patch replaces the CHECK() call with an exit(1) (returing 1 from
main() actually) to prevent sending SIGABRT and collecting crash
reports during the test run. Other minor linter fixes included.

Bug: None
Test: Manually ran the unittests.

Change-Id: I451a3150d72018652b703db7d3b82f4e787c8f43