update_device.py: Handle the package push without 'adb root'.
When using '--file' (and without '--no-push'), it tries to push the
package to /data/ota_package/debug.zip. However, that's not allowed
directly with 'adb push', unless by restarting adb with 'adb root'.

This CL works around the issue by pushing the file to /data/local/tmp
first, then moving it over to the desired place.

Test: update_device.py --file <marlin-ota.zip>
Change-Id: I608284cc90918fc01cf7584569d4d7d4165148ca