Remove duplicates in ab_partitions.txt
ab_partitions.txt is populated by AB_OTA_PARTITIONS which might be
appended in multiple makefiles resulting in duplicates, currently
this will result in payload being larger because it will include the
operations for that partition twice and the client will write that
partition twice when applying that payload.

This patch uses awk in brillo_update_payload when reading
ab_partitions.txt to remove the duplicates.

Test: generate a payload with duplicated entries in ab_partitions.txt
Change-Id: I07bd2401a64fbc77bbdc632901016e3860dee8d9