Save multiple copies of update engine logs
Right now we only have one copy of update engine log; and it's
overwritten if the device reboots or update engine restarts. This CL
save multiple copies of the logs with timestamp in
"/data/misc/update_engine_log/". It also implements a function to
keep the most recent five logs but delete the old ones.

Bug: 72080233
Test: Logs are created with correct permission; and old logs are deleted
when update engine starts.
Change-Id: I9853546bce879e7fbd6380419a85e01256331247