update_engine: remove redundant utils functions
utils::MapContainsKey and utils::SetContainsKey are the same
function and could be unified with more general templating, but
that would just be a reimplementation of base::ContainsKey so we
can just use that instead. Additionally, base::ContainsValue can
be used in place of utils::VectorContainsValue.

BUG=None
TEST=unit tests

Change-Id: Ia8fee68ca8d6a87d232f89380c1e4144b598fad9
Reviewed-on: https://chromium-review.googlesource.com/882455
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Alex Deymo <deymo@google.com>