update_engine: use ScopedTempDir::GetPath
Upstream versions of libchrome have gotten rid of ScopedTempDir's
inline path method because they wanted to introduce a DCHECK that
helps make sure people create the temp directory before using it.
To avoid introducing a header dependency on the logging header
they took this method out of the header so it wouldn't be inlined.

BUG=b:37434548
TEST=unit tests

Change-Id: If1dcc5e43f54ab32fd43da7b939216a4834548d5
Reviewed-on: https://chromium-review.googlesource.com/882543
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>