postinstall: fixed postinst_action_progress is NaN
If no postinstall is defined, total_weight_ is 0.
When total_weight_ is 0, the result of division is NaN.
The NaN is set to postinst_action_progress, and update_engine
sends the NaN for other software(ex. application) by callback.
The software might show the NaN value on screen as decimal
fraction(double).
Update_engine should not send NaN in this case.

Bug: 78199138
Test: manual - Create update.zip with AB_OTA_POSTINSTALL_CONFIG=(null)
      Then, Run ./update_device.py --file update.zip,
      and check log

Change-Id: I1ab5e607bc6598262c47b58b2fe7415195d913a4