update_engine: guarded 740efad for only android
740efad439e57e71d8d459c366edea3a90b19576: Reboot even if a system update is not available.

The above patch removes a part of code in ChromeOS, but the Android still needs
it. That patch has been reverted in AOSP.

This CL guards that piece of code to be built on non-chromeos builds so there
would be no need for reverting it again in android when merging back and forth.

BUG=chromium:815356
TEST=unittests

Change-Id: Ieb8f246d9dba0bd2b7ee7ecc4e3b7823249b3599
Reviewed-on: https://chromium-review.googlesource.com/965265
Commit-Ready: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Sen Jiang <senj@chromium.org>