Fix error code for no update.
PayloadState does not get notified if there's no update, so the error
code will not be updated, move this to UpdateAttempter.

We were using the generic kError for no update, this change added a
new error code kNoUpdate for this, now update_engine_client will show
kNoUpdate instead of kSuccess or a random last error.

Bug: 36946608
Test: update_engine_client --check_for_update --follow
Change-Id: Ie1e94841788437140e0894dc2e73c314a7564108