init: Change the update_engine's respawn rate limit
Currently, we are respawning the update engine daemon if there are 10 crashes
in 5 seconds (the default values). But crbug.com/831815 proved that this rate
limit is quite fast. Even if we fail early, the crash rate will be like 1 crash
per second. Hence, currently the repeated crashes cannot be detected and the
update engine will create a log every second which will eventually fill up the
stateful partition. This patch changes the crash limit rate to 10 failures in 20
seconds. This rate covers the default limit too.

BUG=chromium:831815
TEST=forced the UE to crash and the respawning stopped after 10 times.

Change-Id: I180021c155025262bdb743e1d02ba96eeb09ebad
Reviewed-on: https://chromium-review.googlesource.com/1022737
Commit-Ready: Amin Hassani <ahassani@chromium.org>
Tested-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>