Enable BCJ filters for X86 and ARM ELF binary.
BCJ filter convert relative addresses into absolute addresses to
increase redundancy, but it could make things worse if applied on the
wrong data, so we only enabled it on ELF with supported arch.

Some testing with an ARM system image shows that using the filter could
reduce the compressed size by up to 20% when compressing a single ELF
file using XZ, or increase by up to 1%, on average this will save about
6%.

The saving on final payload size are less significant because bsdiff
will win most of the time in delta payload, and for full payload we
are not splitting the operation by file, so the filter won't apply
most of the time.

Bug: 27878181
Test: generated delta and full payload
Test: update_engine_unittests

Change-Id: I9b6ce1b0b8b0e37cf35a9a6dfb02ebbb8be65109