Support a range of major and minor version.
In DeltaPerformer we set a supported major and minor version and only
accept that perticular version, but in reality we still support older
versions to test backward comatibility.

This CL change the supported major and minor version to a range to
reflect the truth of which version is supported and moved them to
payload_constants.

This is also useful when we have multi payload and want to reuse old
payload for non system partition while updating the system partition
to use a higher minor version.

Bug: 36289531
Test: update_engine_unittests
Change-Id: I49d73d69e64d78dfc2db21446faa98b3579ff995
Merged-In: I49d73d69e64d78dfc2db21446faa98b3579ff995