Use ActionProcessor::IsRunning() to check for ongoing update.
If we started the update but the action processor haven't started yet,
and at this moment we get a call to CancelUpdate(), |ongoing_update_|
will be true but processor_->IsRunning() is false, which will crash by
CHECK(IsRunning()) in ActionProcessor::StopProcessing().

We can simply always use ActionProcessor::IsRunning() to track the state
and remove |ongoing_update_| variable.

Bug: 111354431
Test: start an update and cancel it
Change-Id: I1bcb4ecf62d931e649389062bd6a08834d94a4cc