Skip writing verity if already written.
Computing FEC on device could take up to 3 minutes depending on
partition size, we should skip it if it's already written.
This is similar to how we skip postinstall for postponed OTA, but we
don't require passing additional header here because we can verify the
correctness of the verity data within update_engine itself.

Bug: 28171891
Test: update_engine_unittests

Change-Id: Ie9883e2260d95c05aec169dd1fde12beea0bdade