Consider diff operation overhead.
Storing a diff operation has more overhead over replace operation in
the manifest, we need to store an additional src_sha256_hash which is
32 bytes and not compressible, and also src_extents which could use
anywhere from a few bytes to hundreds of bytes depending on the
number of extents.

We should consider this overhead when deciding whether to use a diff
operation over a replace operation, replace operation should be
prefered if payload size is similar, because they don't rely on
existing data on disk and they could be merged with other replace
operations.

Test: brillo_update_payload generate
Test: brillo_update_payload verify
Change-Id: I850df7baf72fdd1f5b70b22506ebacadb60db58b